Common wait for presentation properties
Wait for transition

Fix storage list spinner key
Unify loader, list and search into list page
