+++ test/e2e/editor/cases/placeholder-settings.js
@@ -43,7 +43,6 @@ var PlaceholderSettingsScenarios = function() {
     describe(' Given a user is adding a new presentation and a few placeholders', function () {
       before(function () {
         presentationsListPage.openNewPresentation();
-        helper.wait(presentationPropertiesModalPage.getPresentationPropertiesModal(), 'Presentation Properties Modal');
         presentationPropertiesModalPage.getCancelButton().click();
 
         workspacePage.getAddPlaceholderButton().click();
