+++ test/e2e/editor/cases/presentation-properties.js
@@ -45,10 +45,6 @@ var PresentationPropertiesScenarios = function() {
       });
 
       describe("Given a user clicks on the presentation properties cog icon", function () {
-        before(function () {
-          helper.wait(presentationPropertiesModalPage.getPresentationPropertiesModal(), 'Presentation Properties Modal');
-
-        });
         it('should open the properties modal', function () {
           expect(presentationPropertiesModalPage.getPresentationPropertiesModal().isDisplayed()).to.eventually.be.true;
         });
