+++ test/e2e/storage/cases/new-folder.js
@@ -55,7 +55,7 @@ var NewFolderScenarios = function() {
       describe("New Folder:", describeNewFolder);
     });
 
-    describe("And he is using Apps Single Folder Selector:",function(){
+    describe("And he is using Apps Single File Selector:",function(){
       before(function () { StorageHelper.setupAppsSingleFileSelector(); });
       describe("New Folder:", describeNewFolder);
     });
