+++ test/e2e/storage/pages/storageHomePage.js
@@ -15,18 +15,7 @@ var StorageHomePage = function() {
   var deleteForeverButton = element(by.id('deleteForeverButton'));
   var confirmDeleteButton = element(by.id('confirmForm')).element(by.buttonText('Delete Forever'));
 
-  var searchInput = element(by.id('storageSelectorSearchInput'));  
-
-  var loader = element(by.xpath('//div[@spinner-key="_rv-global-spinner"]'));
-  var storageFileList = element(by.css('.storage-app .scrollable-list'));
-  var fileListRows = element.all(by.css('#storageFileList > tbody > tr'));
-
   var pendingOperationsPanel = element(by.id('pendingOperationsPanel'));
-
-  this.filterFileList = function(query) {
-    this.getSearchInput().clear();
-    this.getSearchInput().sendKeys(query);
-  }
 
   this.getStorageAppContainer = function() {
     return storageAppContainer;
@@ -80,21 +69,6 @@ var StorageHomePage = function() {
     return confirmDeleteButton;
   }
 
-  this.getSearchInput = function(){
-    return searchInput;
-  }
-
-  this.getFileListRows = function() {
-    return fileListRows;
-  }
-
-  this.getStorageFileList = function() {
-    return storageFileList;
-  }
-
-  this.getLoader = function() {
-    return loader;
-  }
 };
 
 module.exports = StorageHomePage;
