+++ test/e2e/storage/pages/storageSelectorModalPage.js
@@ -11,17 +11,6 @@ var StorageSelectorModalPage = function() {
   var uploadPanel = element(by.css('.upload-panel'));
   var closeButton = element(by.id('closeButton'));
 
-  var searchInput = element(by.id('storageSelectorSearchInput'));  
-
-  var loader = element(by.xpath('//div[@spinner-key="_rv-global-spinner"]'));
-  var storageFileList = element(by.css('.storage-app .scrollable-list'));
-  var fileListRows = element.all(by.css('#storageFileList > tbody > tr'));
-
-  this.filterFileList = function(query) {
-    this.getSearchInput().clear();
-    this.getSearchInput().sendKeys(query);
-  }
-
   this.getStorageSelectorModal = function() {
     return storageSelectorModal;
   };
@@ -58,21 +47,6 @@ var StorageSelectorModalPage = function() {
     return uploadPanel;
   };
 
-  this.getSearchInput = function(){
-    return searchInput;
-  }
-
-  this.getFileListRows = function() {
-    return fileListRows;
-  }
-
-  this.getStorageFileList = function() {
-    return storageFileList;
-  }
-
-  this.getLoader = function() {
-    return loader;
-  }
 };
 
 module.exports = StorageSelectorModalPage;
